Personalized-Multimodal-AI-Tutor-main/requirements.txt
torch
transformers
langchain
langchain-community
langchain-huggingface
faiss-cpu
streamlit>=1.24.0
diffusers
pypdf
accelerate
gpt4all
